requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-filter==2.2.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
oauthlib==3.1.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
